fix(security): Fix regression in adding X-SPINNAKER-ACCOUNTS
